requirements.txt
Flask==1.0.2
dnspython==1.16.0
Flask-RESTful==0.3.7
pymysql==0.9.3
jsonify==0.5
